refactor: Consolidate authentication logic by moving withAuth function to api utility and removing auth.ts file
