+++ src/app/api/profile/update/route.ts
@@ -2,8 +2,7 @@ import { imageFileTypes } from '@/constants/fileTypes';
 import { db } from '@/db/db';
 import { type User, usersTable } from '@/db/schema';
 import { profileSchema } from '@/schema';
-import { handleApiError } from '@/utils/api';
-import { withAuth } from '@/utils/auth';
+import { handleApiError, withAuth } from '@/utils/api';
 import { uploadFile } from '@/utils/uploadFile';
 import { fileValidationConfigs, validateFile } from '@/utils/validateFile';
 import { eq } from 'drizzle-orm';
